+++ gads_lib/__init__.py
@@ -1,6 +1,6 @@
 """Internal library for the gads CLI."""
 
-__version__ = "3.9.1"
+__version__ = "3.10.0"
 
 from .config import (
     PROJECT_ROOT,