+++ generate_token.py
@@ -29,6 +29,7 @@ SCOPES = [
     "https://www.googleapis.com/auth/analytics.readonly",
     "https://www.googleapis.com/auth/analytics.edit",
     "https://www.googleapis.com/auth/webmasters.readonly",
+    "https://www.googleapis.com/auth/datamanager",
 ]
 
 CLIENT_SECRET = CREDENTIALS_DIR / "client_secret.json"